requirements.txt
fastapi
uvicorn
python-socketio
sqlalchemy
duckdb
pydantic
requests
python-dotenv
numpy
pandas
scikit-learn
